+++ client/src/Pages/PlansPage.js
@@ -13,8 +13,21 @@ export const PlanPage = () => {
       const [update, setUpdate] = useState(false)
 
       const auth = useContext(AuthContext)
-      const {request,error,  clearError} = useHttp()
+      const {loading,request,error,  clearError} = useHttp()
       const message = useMessage()  
+    /*useEffect(()=>{
+      const data = request('/api/auth/check', 'GET', null, {
+        Authorization: `Bearer ${auth.token}`
+      })
+      auth.login(data.token, data.userId)},[])
+*/
+       //const updateAuth = useCallback( async()=>{
+       //  const data = request('/api/auth/check', 'GET', null, {
+       //    Authorization: `Bearer ${auth.token}`
+       //  })
+       //  auth.login(data.token, data.userId)
+       //})
+
       const fetchPlans = useCallback(async () => {
         try {
           const fetched = await request('/api/todo', 'GET', null, {
@@ -53,6 +66,7 @@ export const PlanPage = () => {
 
     function updateData(id){
       setUpdate(id)
+      console.log(loading)
     }
   
       function removeToDo(id){
@@ -70,49 +84,39 @@ export const PlanPage = () => {
     settodos(todos.filter(todo => todo._id !== id))
   }
 
-  if (todos.length<1) // загрузка окна // багается когда запускается сервер и заранее окно открыто *это не фикситься*
+  if (todos.length<1)
   {
     return (
-      <div style={{backgroundColor: "rgb(240, 240, 240)"}}>
-        <div style={{
-                display:'flex',
-                flexDirection:'column',
-                margin:'0 auto',
-                width:'100%',
-                maxWidth:'960px',
-                height:'100%',
-                minHeight:'830px',
-                alignItems:'center',
-                justifyContent:'center',
-                paddingTop: '2rem',
-                paddingBottom: '2rem'
-            }}>
-            <Loader type="Circles" color="#00BFFF" height={80} width={80}/>
-          </div>
-        </div>
-
-  )
-}
-return (
-    <Context.Provider value={{removeToDo,updateData}}>
-      <div style={{backgroundColor: "rgb(240, 240, 240)"}}>
-        <div style={{
-                display:'flex',
-                flexDirection:'column',
-                margin:'0 auto',
-                width:'100%',
-                maxWidth:'960px',
-                height:'100%',
-                minHeight:'830px',
-                alignItems:'center',
-                justifyContent:'center',
-                paddingTop: '2rem',
-                paddingBottom: '2rem'
-            }}>
-            <ModalPlan></ModalPlan>      
-            <Todolist todos={todos} onToggle={toggleToDo}></Todolist>
+<div className='jarallax-container-0' style= {{backgroundColor: "rgb(240, 240, 240)"}}>
+  <div className = "row">
+    <div className="col s8 offset-s9" style={{paddingTop: '20rem', paddingBottom:'20rem'}}>
+      <div className=" container">
+        <div className="row justify-content-center">
+          <Loader type="Circles" color="#00BFFF" height={80} width={80}/>
         </div>
       </div>
-    </Context.Provider>
-  )
+    </div>
+  </div>
+</div>)
+  }
+return (
+<Context.Provider value={{removeToDo,updateData}}>
+  <div className='jarallax-container-0' style= {{backgroundColor: "rgb(240, 240, 240)"}}>
+    <div className='row justify-content-center' style={{paddingTop:"1rem", paddingBottom:'1rem'}}>
+      <ModalPlan></ModalPlan>
+    </div>
+        <div className = "row">
+          <div className="col s8 offset-s9" style={{paddingBottom:'1rem'}}>
+            <div className=" container">
+              <div className="row justify-content-center">
+                  <div className="col-12 col-lg-8">
+                    <Todolist todos={todos} onToggle={toggleToDo}></Todolist>
+                  </div>
+              </div>
+            </div>
+          </div>
+        </div>
+  </div>
+</Context.Provider>
+)
 }
