+++ client/src/components/ModalPlan.js
@@ -37,6 +37,7 @@ export const ModalPlan = () => {
             Authorization: `Bearer ${auth.token}`
           })
           setShow(false)
+          console.log(data.message)
           message(data.message)
           form.title=''
           form.describe=''
@@ -46,8 +47,9 @@ export const ModalPlan = () => {
 
     
 return (
-  <div>
-        <Button variant="primary" style={{marginBottom:'1rem'}} onClick={handleShow}>Добавить план</Button>
+<div className='jarallax-container-0' style= {{backgroundColor: "rgb(240, 240, 240)"}}>
+    <div className='row justify-content-center' style={{paddingTop:"2rem"}}>
+        <Button variant="primary" onClick={handleShow}>Добавить план</Button>
 
         <Modal show={show} onHide={handleClose}> 
             <Modal.Header style={{justifyContent:'center'}}>
@@ -93,7 +95,8 @@ return (
                 </Button>
           </Modal.Footer>
         </Modal>
-        </div>
+    </div>
+</div>
 )
 }
 export default ModalPlan
