+++ client/src/components/Todoitem.js
@@ -1,9 +1,25 @@
 import PropTypes from 'prop-types'
 import React, {useState,useContext} from 'react';
 import Context from '../context/context'
-import ModalTodo from './ModalTodo';
+import { Button, Modal } from 'react-bootstrap';
 
+const style={
+    li:
+    {
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        padding: '.5rem .1rem',
+        border: '1px solid #ccc',
+        borderRadius:'4px',
+        marginBottom: '.5rem'
+    },
 
+    input:
+    {
+        marginRight:'1rem',
+    }
+}   
 
 
 function TodoItem({todo, index, onChange})  {
@@ -12,51 +28,41 @@ function TodoItem({todo, index, onChange})  {
     if (todo.completed) {
         classes.push('done')
     }
-    const [info, setInfo] = useState(false)
-    const fuck= () =>
+    
+    const [show, setShow] = useState(false);
+    const handleClose = () => 
     {
-        setInfo(true)
-    }
-    const [hovered, setHovered] = useState(false)
+        setShow(false)
+        
+    };
+  const handleShow = () => setShow(true);
+
 
     return (
-    <div onMouseMove={() => setHovered(true)} onMouseLeave={()=>setHovered(false)}>
-        {hovered ?
-        (
-        <li style={{display:'flex',flexDirection:'row',border:'1px solid blue', borderRadius:'4px', marginBottom:'3px' ,alignItems: 'center', justifyContent:'space-between', cursor:'pointer', backgroundColor:'whitesmoke'}}>
-            <div style={{display:'flex'}}>
+        <li style={style.li}>
+            <span className={classes.join('')}>
                 <input type='checkbox' 
                 checked={todo.completed}
-                style={{margin:'1rem'}}
-                onChange={()=> onChange(todo._id)}/> 
-                <span className={classes.join('')} style={{wordBreak:'break-word'}}>
-                    <div style={{display:'flex',marginTop:'0.6rem'}} onClick={fuck}>
-                        <p style={{paddingRight:'2px'}}><strong>{index+1}</strong></p>
-                        <p>{todo.title}</p>
-                    </div>
-                </span>
-            </div>
-            <button style={{borderRadius:'10px'}} onClick = {removeToDo.bind(null,todo._id)}>✘</button>
+                style={style.input}
+                onChange={()=> onChange(todo._id)}> 
+                </input>
+                
+                <strong  onClick={handleShow}>{index+1} </strong>
+                <a  onClick={handleShow}>{todo.title}</a>
+            </span>
+            <Modal show={show} onHide={handleClose}>
+        <Modal.Header closeButton>
+          <Modal.Title>{todo.title}</Modal.Title>
+        </Modal.Header>
+        {todo.describe.length>1 ?<Modal.Body><span style={{wordBreak:'break-word'}}>{todo.describe}</span></Modal.Body>:<Modal.Body>Тут ничего нет</Modal.Body>}
+        <Modal.Footer>
+        <strong>Чтобы закрыть нажмите ESC</strong>
+        </Modal.Footer>
+      </Modal>
+        <button 
+        onClick = {removeToDo.bind(null,todo._id)}>
+        ✘</button>
         </li>
-        ) : 
-        (
-        <li style={{display:'flex',flexDirection:'row',border:'1px solid black', borderRadius:'4px', marginBottom:'3px' ,alignItems: 'center', justifyContent:'space-between', cursor:'pointer'}}>
-            <div style={{display:'flex'}}>
-                <input type='checkbox' 
-                checked={todo.completed}
-                style={{margin:'1rem'}}
-                onChange={()=> onChange(todo._id)}/> 
-                <span className={classes.join('')} style={{wordBreak:'break-word'}}>
-                    <div style={{display:'flex',marginTop:'0.6rem'}}>
-                        <p style={{paddingRight:'2px'}}><strong>{index+1}</strong></p>
-                        <p>{todo.title}</p>
-                    </div>
-                </span>
-            </div>
-            <button style={{borderRadius:'10px'}}onClick = {removeToDo.bind(null,todo._id)}>✘</button>
-        </li>)}
-        <ModalTodo item={info} todo={todo}></ModalTodo>
-    </div>
     )
 }
